organize the Request classes for tests
There was a `tests/Fixtures/Requests` directory, made sense to move the Request classes there.
