+++ tests/Fixtures/Controllers/ValidationController.php
@@ -9,6 +9,7 @@ use Tempest\Http\Responses\Ok;
 use Tempest\Http\Responses\Redirect;
 use Tempest\Router\Get;
 use Tempest\Router\Post;
+use Tests\Tempest\Fixtures\Requests\ValidationRequest;
 
 use function Tempest\uri;
 
@@ -21,7 +22,7 @@ final readonly class ValidationController
     }
 
     #[Post('/test-validation-responses')]
-    public function store(RequestForValidationController $request): Response // @mago-expect best-practices/no-unused-parameter
+    public function store(ValidationRequest $request): Response // @mago-expect best-practices/no-unused-parameter
     {
         return new Redirect(uri([self::class, 'get']));
     }
