+++ tests/Fixtures/Requests/ValidationRequest.php
@@ -2,17 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Tests\Tempest\Fixtures\Controllers;
+namespace Tests\Tempest\Fixtures\Requests;
 
 use Tempest\Http\IsRequest;
 use Tempest\Http\Request;
 use Tempest\Validation\Rules\Between;
 
-final class RequestForValidationController implements Request
+final class ValidationRequest implements Request
 {
     use IsRequest;
-
-    public RequestItemForValidationController $item;
 
     #[Between(min: 1, max: 10)]
     public int $number;
